Update: Reorganize types
- Move core types out of global.d.ts
- Fix type errors
