+++ packages/core/package.json
@@ -51,7 +51,7 @@
   "devDependencies": {
     "@babel/preset-typescript": "^7.8.3",
     "@ember/optional-features": "^1.3.0",
-    "@types/ember": "^3.1.2",
+    "@types/ember": "^3.16.0",
     "@types/ember-data": "^3.16.2",
     "@types/ember-data__model": "^3.16.0",
     "@types/ember-data__store": "^3.16.0",
