+++ packages/core/types/ember-cp-validations/index.d.ts
@@ -1,0 +1,8 @@
+declare module 'ember-cp-validations' {
+  type EmberCPValidationsMixin = {
+    validations: TODO;
+  };
+  function buildValidations(validation: TODO): EmberCPValidationsMixin;
+  function validator(name: string, options?: TODO): TODO;
+  export { buildValidations, validator };
+}
