+++ packages/reports/package.json
@@ -42,7 +42,7 @@
     "ember-cli-moment-shim": "^3.8.0",
     "ember-cli-node-assets": "0.2.2",
     "ember-cli-shims": "^1.2.0",
-    "ember-cli-typescript": "^3.1.3",
+    "ember-cli-typescript": "^3.1.4",
     "ember-collection": "1.0.0-alpha.9",
     "ember-composable-helpers": "^3.1.1",
     "ember-concurrency": "1.1.2",
@@ -72,6 +72,8 @@
     "@glimmer/component": "^1.0.0",
     "@types/ember": "^3.16.0",
     "@types/ember-data": "^3.16.2",
+    "@types/ember-data__model": "^3.16.0",
+    "@types/ember-data__store": "^3.16.0",
     "@types/ember-qunit": "^3.4.9",
     "@types/ember__test-helpers": "^1.7.2",
     "@types/qunit": "^2.9.1",
@@ -125,7 +127,7 @@
     "navi-notifications": "0.2.0",
     "npm-run-all": "^4.1.5",
     "qunit-dom": "^1.2.0",
-    "typescript": "^3.9.7"
+    "typescript": "^4.0.2"
   },
   "engines": {
     "node": "10.* || >= 12"
