+++ packages/reports/types/ember-data/types/registries/transform.d.ts
@@ -1,0 +1,3 @@
+export default interface TransformRegistry {
+  [key: string]: any;
+}
